+++ client/src/utils/convertToKoreanDate.ts
@@ -1,0 +1,11 @@
+/** 12345를 집어넣으면 "+12,345" 반환, -98765를 집어넣으면 "-98,765" 반환 */
+export function convertToKoreanDate(date: Date) {
+  const year = date.getFullYear();
+  const month = date.getMonth() + 1;
+  const day = date.getDate();
+  const hours = date.getHours().toString().padStart(2, '0'); // 한 자릿수라면 앞에 0 추가
+  const minutes = date.getMinutes().toString().padStart(2, '0'); // 한 자릿수라면 앞에 0 추가
+
+  const formattedDate = `${year}년 ${month}월 ${day}일 ${hours}:${minutes}`;
+  return formattedDate;
+}
